// src/error/custom-toaster/enum.ts
export enum CustomToastErrorMessages {
  defaultTitle = "Ops...",
  commentErrorMessage = "Parece que houve um erro ao criar o seu comentário, por favor, tente novamente.",
  postErrorMessage = "Parece que algum erro aconteceu durante o processo de criação do seu post, pedimos desculpa pelo ocorrido. Por favor, tente novamente.",
  wrongPassword = "Nome de usuário ou senha incorreta",
  wrongAuthRequest = "Parece que ocorreu um erro no servidor, por favor, tente novamente",
}

export enum CustomToastProps {
  defaultDuration = 5000,
}